requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-environ==0.8.1
django-heroku==0.3.1
django-recaptcha==2.0.6
environ==1.0
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
